+++ tools/code-formatter/black.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+SRC=samples/main.py
+
+echo "----------- yapf -----------"
+black --diff ${SRC}