+++ tools/code-formatter/indent.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+SRC=samples/main.c
+
+echo "------------- GNU style -------------"
+indent -st ${SRC}
+
+echo "------------- Linux style -------------"
+indent --linux-style -st ${SRC}