+++ tools/code-formatter/samples/main.py
@@ -1,0 +1,10 @@
+#!/bin/env python3
+
+import os
+
+
+def hello(name):
+    print("Hello World! %s" % (name))
+
+
+hello("OSTools")