+++ tools/code-formatter/yapf.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+SRC=samples/main.py
+
+echo "----------- yapf -----------"
+yapf ${SRC}