+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- tools: add code-formatter from ostools;
 - template: add gcc.mk;
 - cuda/samples: add graph.cu;
 - gcc: add version.sh;